Squashed commit of the following:

commit ec308cd55b2e2e32ae4db3eb969b5fd38e7bb63f
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Mon Apr 16 16:49:26 2018 -0400

    stream_tcp: switch to splitter max

commit 783c405bb721e48352db0a9e00946722c037231a
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Sat Apr 14 07:46:20 2018 -0400

    spell check: fix memeory and indicies typos

commit 9aa84591ec853873196ca1bb5dc85fddd043081a
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Fri Apr 13 21:41:02 2018 -0400

    build: fix various warnings:

    file_decomp_pdf.cc: unreachable code return
    decompress/file_decomp_pdf.cc: implicit fallthrough
    stream.cc: warning: potential memory leaks
    memory_manager.cc: dangling references

commit f07b179baae636ec9be6307fe3bbb0295aeba387
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Thu Apr 12 09:27:50 2018 -0400

    stream: be selective about flow creation for scans

commit cdf528691f7aaeebb40a0b686ae46b7ff4651579
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Wed Apr 11 21:25:31 2018 -0400

    http_inspect: patch around regression failures until a permanent solution is implemented

commit 7f57541f3e4d1c1aaf44807c358b858f9ede043c
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Fri Apr 6 17:17:56 2018 -0400

    stream_tcp: tweak seglist cursor handling

commit 417992054439d839d3651fc5adf1b08d836cd1d4
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Wed Apr 11 07:41:35 2018 -0400

    port_scan: include open ports with alerts instead of separate

commit cf2e7d2ac36b3cf70ee8c7a55f932029755ddccb
Author: Russ Combs <rucombs@cisco.com>
Date:   Sat Mar 31 12:41:53 2018 -0400

    stream: remove usused ignore_any_rules from tcp and udp

commit f5843c4d2ad6a439003afde1830c41900bc492bf
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Thu Apr 5 17:00:27 2018 -0400

    stream: respect tcp require_3whs

commit c9bb1eb11d3060515bc26e8a393a1c963f47ae9c
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Thu Apr 5 11:44:15 2018 -0400

    stream: refactor flow control for new PktTypes

commit c9c5d239de87e9aa45925fde906d842f3ad2c33d
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Tue Apr 10 11:12:06 2018 -0400

    search_engine: revert default detect_raw_tcp to false

commit f0556f01cf98bfd8fef7494177ae6b33fde52b38
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Thu Apr 5 09:57:27 2018 -0400

    framework: remove bogus PktType for ARP and just use proto bits instead

commit f84ee2cb693c94e243180b2f3a715cc564d2d329
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Thu Apr 5 09:31:53 2018 -0400

    framework: align PktType and proto bits

commit f1f54c0fd3188460b1cb2e203cfb3c4789ed8e8a
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Mon Apr 2 06:14:05 2018 -0400

    memory: disable with -DNO_MEM_MGR

commit 6d070a4e8f17ae96f519cb833e356aaa7c3e6145
Author: Russ Combs (rucombs) <rucombs@cisco.com>
Date:   Mon Apr 2 06:13:42 2018 -0400

    profile: disable with -DNO_PROFILER

commit a7ddf1395e9c9e69eb2b5d72077c40cb5e7aac19
Author: Russ Combs <rucombs@cisco.com>
Date:   Sat Mar 31 13:04:24 2018 -0400

    zhash: cleanup cruftiness

+++ src/framework/codec.h
@@ -126,7 +126,7 @@ struct CodecData
     /* Reset before each decode of packet begins */
 
     /*  Codec specific fields.  These fields are only relevant to codecs. */
-    uint16_t proto_bits;    /* protocols contained within this packet
+    uint32_t proto_bits;    /* protocols contained within this packet
                                  -- will be propogated to Snort++ Packet struct*/
     uint16_t codec_flags;   /* flags used while decoding */
     uint8_t ip_layer_cnt;
